# cloudnet-rest-module/src/main/resources/documentation/swagger.yaml
openapi: 3.1.0
info:
  description: >
    CloudNet's rest API version 3 is the next generation in CloudNet's rest
    history.
    
    
    In order to access rest API v3 you will have to authenticate yourself using basic auth.
    To create a basic auth user use the `rest user create <id> <password>` command on your cloudnet node.
    After authenticating with basic auth on `/auth` you will get a bearer jwt access token that must be used for further
    requests to the api. For further information about the response take a look at the `/auth` endpoint.
    
    MENTION CONFIG HERE
    MENTION CASE-SENS HERE
  version: '3'
  title: CloudNet
  contact:
    name: CloudNet
    url: https://cloudnetservice.eu
  license:
    name: Apache license, version 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: '{protocol}://{host}:{port}/api/v3/'
    variables:
      protocol:
        enum:
          - http
          - https
        default: https
      host:
        default: 127.0.0.1
      port:
        default: '2812'
security:
  - bearerAuth: [ ]
...
paths:
...
  /cluster:
    get:
      tags:
        - Cluster
      summary: Lists all nodes
      description: |
        Lists all nodes which are registered in the node. This includes nodes
        which are currently not connected to the node. If they are not connected
        available will be `false` and the `nodeInfoSnapshot` will be null.

        One of the following scopes is needed to execute the request:
         - `cloudnet_rest:cluster_read`
         - `cloudnet_rest:cluster_node_list`
      operationId: cluster
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NodeServer'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
    put:
      requestBody:
        required: true
        description: The new node configuration entry.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NetworkClusterNode'
      tags:
        - Cluster
      summary: Updates a node
      description: |
        Changes the configuration of a specific node in the local node's config
        file. The update will be pushed live to the node cluster settings. That
        does not mean that the node tries to connect to the updated node.

        One of the following scopes is needed to execute the request:
         - `cloudnet_rest:cluster_write`
         - `cloudnet_rest:cluster_node_update`
      operationId: clusterUpdate
      responses:
        '204':
          description: The node was updated successfully
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
    post:
      requestBody:
        required: true
        description: The node configuration entry.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NetworkClusterNode'
      tags:
        - Cluster
      summary: Creates a new node
      description: |
        Creates a node in the node's local configuration and updates it live to
        the node cluster settings. That does not mean that the node tries to
        connect to the newly registered node.

        One of the following scopes is needed to execute the request:
         - `cloudnet_rest:cluster_write`
         - `cloudnet_rest:cluster_node_create`
      operationId: clusterCreate
      responses:
        '201':
          description: The new node was created successfully
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
...
  /service:
    get:
      tags:
        - Services
      summary: Lists all known services
      description: |
        One of the following scopes is needed to execute the request:
          - `cloudnet_rest:service_read`

          - `cloudnet_rest:service_list`
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ServiceInfoSnapshot'
          description: Successful listing of services.
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          description: There is no service with the provided name / id.
          $ref: '#/components/responses/Problem'
  /service/{identifier}:
    parameters:
      - name: identifier
        in: path
        required: true
        description: The name or unique id of the service
        schema:
          type: string
    get:
      tags:
        - Services
      summary: Shows a specific service if it's exists
      description: |
        One of the following scopes is needed to execute the request:
          - `cloudnet_rest:service_read`

          - `cloudnet_rest:service_get`
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceInfoSnapshot'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          description: There is no service with the provided name / id.
          $ref: '#/components/responses/Problem'
    delete:
      tags:
        - Services
      summary: Deletes a service
      description: |
        One of the following scopes is needed to execute the request:
          - `cloudnet_rest:service_write`
        
          - `cloudnet_rest:service_delete`
      responses:
        '204':
          description: The service was deleted
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          description: There is no service with the provided name / id.
          $ref: '#/components/responses/Problem'
  /service/{identifier}/deleteFiles:
    parameters:
      - name: identifier
        in: path
        required: true
        description: The name or unique id of the service
        schema:
          type: string
    delete:
      tags:
        - Services
      summary: Deletes all files of a service
      description: |
        One of the following scopes is needed to execute the request:
          - `cloudnet_rest:service_write`
        
          - `cloudnet_rest:service_delete_files`
        
        WARNING: This will delete files of static services too.
      responses:
        '204':
          description: The services files were deleted
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          description: There is no service with the provided name / id.
          $ref: '#/components/responses/Problem'
  /service/{identifier}/lifecycle:
    parameters:
      - name: identifier
        in: path
        required: true
        description: The name or unique id of the service
        schema:
          type: string
      - name: target
        in: query
        required: true
        description: The target service lifecycle
        schema:
          type: string
          enum:
            - start
            - stop
            - restart
    patch:
      tags:
        - Services
      summary: Updates the lifecycle of a service
      description: |
        One of the following scopes is needed to execute the request:
          - `cloudnet_rest:service_write`

          - `cloudnet_rest:service_lifecycle`
      responses:
        '204':
          description: The service lifecycle was changed
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          description: There is no service with the provided name / id.
          $ref: '#/components/responses/Problem'
  /service/{identifier}/include:
    parameters:
      - name: identifier
        in: path
        required: true
        description: The name or unique id of the service
        schema:
          type: string
      - name: type
        in: query
        required: true
        description: The type of inclusion to include
        schema:
          type: string
          enum:
            - templates
            - inclusions
    post:
      tags:
        - Services
      summary: Includes a specific type of inclusion added to a service
      description: |
        One of the following scopes is needed to execute the request:
          - `cloudnet_rest:service_write`

          - `cloudnet_rest:service_include`
      responses:
        '204':
          description: All waiting components of the given type were included.
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          description: There is no service with the provided name / id.
          $ref: '#/components/responses/Problem'
  /service/{identifier}/deployResources:
    parameters:
      - name: identifier
        in: path
        required: true
        description: The name or unique id of the service
        schema:
          type: string
      - name: remove
        in: query
        required: false
        description: If the deployments should be removed afterwards
        schema:
          type: boolean
          default: true
    post:
      tags:
        - Services
      summary: Deploys all waiting deployments of a service
      description: |
        One of the following scopes is needed to execute the request:
          - `cloudnet_rest:service_write`

          - `cloudnet_rest:service_deploy_resources`
      responses:
        '204':
          description: All deployments were executed
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          description: There is no service with the provided name / id.
          $ref: '#/components/responses/Problem'
  /service/{identifier}/logLines:
    parameters:
      - name: identifier
        in: path
        required: true
        description: The name or unique id of the service
        schema:
          type: string
    get:
      tags:
        - Services
      summary: Get the cached log lines of a service
      description: |
        One of the following scopes is needed to execute the request:
          - `cloudnet_rest:service_read`

          - `cloudnet_rest:service_log_lines`
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                example:
                  - Line1
                  - Line2
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          description: There is no service with the provided name / id.
          $ref: '#/components/responses/Problem'
...
  /service/create/serviceConfig:
    post:
      parameters:
        - name: start
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: Whether the created service should be started
      tags:
        - Services
      summary: Creates a new service based on a service configuration
      description: |
        One of the following scopes is needed to execute the request:
          - `cloudnet_rest:service_write`

          - `cloudnet_rest:service_create_service_config`
      requestBody:
        required: true
        description: >
          The service configuration to create the new service from
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceConfiguration'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceCreateResult'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
  /service/create/task:
    post:
      parameters:
        - name: start
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: Whether the created service should be started
      tags:
        - Services
      summary: Creates a new service based on a service task
      description: |
        One of the following scopes is needed to execute the request:
          - `cloudnet_rest:service_write`

          - `cloudnet_rest:service_create_task`
      requestBody:
        required: true
        description: >
          The service task to create the new service from
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceTask'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceCreateResult'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
  /service/create/taskName:
    post:
      tags:
        - Services
      summary: Creates a new service
      description: |
        One of the following scopes is needed to execute the request:
          - `cloudnet_rest:service_write`

          - `cloudnet_rest:service_create_task_name`
      requestBody:
        required: true
        content:
          application/json:
            schema:
              taskName:
                type: string
                example: Lobby
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceCreateResult'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    basicAuth:
      type: http
      scheme: basic
  responses:
    Problem:
      description: The action couldn't be exececuted
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
...
    NodeServer:
      type: object
      properties:
        node:
          $ref: '#/components/schemas/NetworkClusterNode'
        state:
          type: string
          enum:
            - UNAVAILABLE
            - SYNCING
            - READY
            - DISCONNECTED
        head:
          type: boolean
          description: If the node server is the current head node.
        local:
          type: boolean
          description: If the node server is the node that handled the request.
        nodeInfoSnapshot:
          $ref: '#/components/schemas/NodeInfoSnapshot'
...